Solution to Check MSB of a Number

+++ Check MSB of a Number@@67882a731b4392d8fc1577e9/code.c
@@ -1,9 +1,12 @@
 // Your code here...
+//If MSB = 0, the number is positive (or zero).
+//If MSB = 1, the number is negative.
+
 #include<stdio.h>
 int main(){
 
-    double x;
-    scanf("%lf", &x);
+    long x;
+    scanf("%ld", &x);
 
     if(x%2==0){
         printf("Not Set");
